refactor(api): git branches 라우트를 전용 모듈로 추출
10개 라우트를 api/git/branches.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 동일.

불연속 두 구역을 한 파일로 합쳤다: Branch Endpoints 6개(원본 660~841) +
Branch Protection Rule Endpoints 4개(원본 1388~1577). 402줄로 800줄 한도
내이므로 branch_protection.py 분리는 하지 않았다.

테스트 패치 타깃 21곳 갱신 (tests/backend/api/test_api_git_prune.py):
  "api.git._legacy.*" → "api.git.branches.*"
  (get_git_service_for_project 7 · get_github_service 7 · _get_db_session 7)
패치 지점은 헬퍼가 정의된 모듈이 아니라 핸들러가 사는 모듈이다. 핸들러가
from ._shared import 로 이름을 자기 네임스페이스에 바인딩하므로, 핸들러
prune_merged_branches 가 이동한 이상 _legacy 를 겨냥한 패치는 무효가 된다.
Task 4·5 에서는 헬퍼만 옮기고 핸들러가 _legacy 에 남아 이 갱신이 불필요했다.
draft-commits 핸들러를 겨냥한 test_llm_usage_instrumentation.py 의 2곳은
Task 7 이 옮길 때까지 _legacy 그대로 둔다.

_shared.py 승격 2건:
- _get_db_session — 계획된 승격 (구역내 6회, 잔여 _legacy 10회)
- get_mr_service_for_project — 계획 외. 브리프 헬퍼 표가 이름이 비슷한 별개
  함수 get_merge_service_for_project(구역내 0회)를 적어 누락했으나, 구역 A의
  delete_branch 가 실제로 호출하는 것은 get_mr_service_for_project 였다.
  레시피 R3(핸들러가 쓰는 _legacy 고유 이름은 _shared 로 승격)에 따라 승격.
  대안인 branches → _legacy import 는 "의존은 항상 한 방향" 불변식을 깬다.
  본문이 쓰는 이름이 모두 _shared 에 이미 있어 상단 import 추가는 0건이다.
두 함수 모두 본문 무변경(무타입 db_session 시그니처 포함).

import 정리: models.git 에서 11개 심볼 + 고아 그룹 주석 3개 제거(잔여 참조
0 실측). can_merge_to_branch 는 merge 도메인이 3회 쓰므로 잔류.
get_github_service 는 구역 A 가 유일한 소비자였어서 함께 고아가 되어 제거.

게이트: ruff check/format 통과, mypy error 0, 라우트 63행 불변,
test_api_git_prune.py 7/7 통과, 전체 1340 passed(기준선 동일 — 유일 실패는
알려진 로컬 .env 플레이크 test_embedding_model_consistency).
